refactor(ui): move test-only quantity/field-manager helpers to tests
